test: stream unit tests
