+++ pkg/mycarehub/infrastructure/services/getstream/stream_test.go
@@ -2,58 +2,17 @@ package getstream_test
 
 import (
 	"context"
-	"reflect"
+	"fmt"
 	"testing"
+	"time"
 
 	stream "github.com/GetStream/stream-chat-go/v5"
+	"github.com/brianvoe/gofakeit"
 	"github.com/google/uuid"
 	"github.com/savannahghi/mycarehub/pkg/mycarehub/domain"
 	"github.com/savannahghi/mycarehub/pkg/mycarehub/infrastructure/services/getstream"
+	mockGetstream "github.com/savannahghi/mycarehub/pkg/mycarehub/infrastructure/services/getstream/mock"
 )
-
-func TestGetStreamClient_CreateUserGetStreamToken(t *testing.T) {
-	type args struct {
-		ctx    context.Context
-		userID string
-	}
-	tests := []struct {
-		name    string
-		args    args
-		want    string
-		wantErr bool
-	}{
-		{
-			name: "Happy Case - Successfully generate a user token",
-			args: args{
-				ctx:    context.Background(),
-				userID: "fe9a8f7c-f8f9-4f0c-b8b1-f8b8f8b8f8b8",
-			},
-			wantErr: false,
-		},
-		{
-			name: "Sad Case - Fail to generate token",
-			args: args{
-				ctx: context.Background(),
-			},
-			wantErr: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			getStream := getstream.NewServiceGetStream(streamClient)
-			got, err := getStream.CreateGetStreamUserToken(tt.args.ctx, tt.args.userID)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("GetStreamClient.CreateGetStreamUserToken() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-
-			if !tt.wantErr && got == "" {
-				t.Errorf("expected a response but got: %v", got)
-				return
-			}
-		})
-	}
-}
 
 func TestChatClient_ListGetStreamUsers(t *testing.T) {
 	type args struct {
@@ -88,9 +47,22 @@ func TestChatClient_ListGetStreamUsers(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			getStream := getstream.NewServiceGetStream(streamClient)
+			fakeClient := mockGetstream.NewStreamClientMock()
+			g := getstream.NewServiceGetStream(fakeClient)
 
-			got, err := getStream.ListGetStreamUsers(tt.args.ctx, tt.args.input)
+			if tt.name == "Happy Case - Successfully list get stream users" {
+				fakeClient.MockQueryUsersFn = func(ctx context.Context, q *stream.QueryOption, sorters ...*stream.SortOption) (*stream.QueryUsersResponse, error) {
+					return &stream.QueryUsersResponse{}, nil
+				}
+			}
+
+			if tt.name == "Sad Case - Fail to get users" {
+				fakeClient.MockQueryUsersFn = func(ctx context.Context, q *stream.QueryOption, sorters ...*stream.SortOption) (*stream.QueryUsersResponse, error) {
+					return nil, fmt.Errorf("fail to query users")
+				}
+			}
+
+			got, err := g.ListGetStreamUsers(tt.args.ctx, tt.args.input)
 			if (err != nil) != tt.wantErr {
 				t.Errorf("ChatClient.ListGetStreamUsers() error = %v, wantErr %v", err, tt.wantErr)
 				return
@@ -102,78 +74,6 @@ func TestChatClient_ListGetStreamUsers(t *testing.T) {
 		})
 	}
 }
-
-func TestChatClient_CreateChannel(t *testing.T) {
-	g := getstream.NewServiceGetStream(streamClient)
-
-	ctx := context.Background()
-	testChannelID := "streamTestChannel"
-
-	type args struct {
-		ctx      context.Context
-		chanType string
-		chanID   string
-		userID   string
-		data     map[string]interface{}
-	}
-	tests := []struct {
-		name    string
-		args    args
-		want    *stream.CreateChannelResponse
-		wantErr bool
-	}{
-		{
-			name: "Happy case - Create channel",
-			args: args{
-				ctx:      ctx,
-				chanType: "messaging",
-				chanID:   testChannelID,
-				userID:   userToAddToNewChannelID,
-				data: map[string]interface{}{
-					"age": map[string]interface{}{
-						"lowerBound": 10,
-						"upperBound": 20,
-					},
-				},
-			},
-			wantErr: false,
-		},
-		{
-			name: "Sad case - empty channel id",
-			args: args{
-				ctx:      ctx,
-				chanType: "test",
-				chanID:   "",
-				userID:   userToAddToNewChannelID,
-				data:     nil,
-			},
-			wantErr: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := g.CreateChannel(tt.args.ctx, tt.args.chanType, tt.args.chanID, tt.args.userID, tt.args.data)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("ChatClient.CreateChannel() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if !tt.wantErr && got == nil {
-				t.Errorf("expected a response but got: %v", got)
-				return
-			}
-		})
-	}
-
-	_, err := c.RemoveMembersFromCommunity(ctx, testChannelID, []string{userToAddToNewChannelID}, nil)
-	if err != nil {
-		t.Errorf("ChatClient.DeleteUsers() error = %v", err)
-	}
-	_, err = g.DeleteChannels(ctx, []string{"messaging:" + testChannelID}, true)
-	if err != nil {
-		t.Errorf("ChatClient.DeleteChannel() error = %v", err)
-	}
-}
-
 func TestChatClient_ListGetStreamChannels(t *testing.T) {
 	type args struct {
 		ctx   context.Context
@@ -199,9 +99,18 @@ func TestChatClient_ListGetStreamChannels(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			getStream := getstream.NewServiceGetStream(streamClient)
+			fakeClient := mockGetstream.NewStreamClientMock()
+			g := getstream.NewServiceGetStream(fakeClient)
 
-			got, err := getStream.ListGetStreamChannels(tt.args.ctx, tt.args.input)
+			if tt.name == "Happy Case - Successfully list get stream channels" {
+				fakeClient.MockQueryChannelsFn = func(ctx context.Context, q *stream.QueryOption, sort ...*stream.SortOption) (*stream.QueryChannelsResponse, error) {
+					return &stream.QueryChannelsResponse{
+						Channels: []*stream.Channel{{ID: gofakeit.UUID()}},
+					}, nil
+				}
+			}
+
+			got, err := g.ListGetStreamChannels(tt.args.ctx, tt.args.input)
 			if (err != nil) != tt.wantErr {
 				t.Errorf("ChatClient.ListGetStreamChannels() error = %v, wantErr %v", err, tt.wantErr)
 				return
@@ -219,21 +128,30 @@ func TestChatClient_GetChannel(t *testing.T) {
 		ctx       context.Context
 		channelID string
 	}
+
 	tests := []struct {
 		name    string
 		args    args
 		wantErr bool
 	}{
 		{
-			name: "Happy case - successfully retrieve a getstream channel",
+			name: "Happy case - retrieve channel by id",
 			args: args{
 				ctx:       context.Background(),
-				channelID: channelID,
+				channelID: gofakeit.UUID(),
 			},
 			wantErr: false,
 		},
 		{
-			name: "Sad case - channel does not exist",
+			name: "Sad case - multiple channels found",
+			args: args{
+				ctx:       context.Background(),
+				channelID: gofakeit.UUID(),
+			},
+			wantErr: true,
+		},
+		{
+			name: "Sad case - non existent channel",
 			args: args{
 				ctx:       context.Background(),
 				channelID: "no-existent-channel",
@@ -241,10 +159,10 @@ func TestChatClient_GetChannel(t *testing.T) {
 			wantErr: true,
 		},
 		{
-			name: "Sad case - invalid channel id",
+			name: "Sad case - error retrieving channel",
 			args: args{
 				ctx:       context.Background(),
-				channelID: "",
+				channelID: "no-existent-channel",
 			},
 			wantErr: true,
 		},
@@ -252,8 +170,40 @@ func TestChatClient_GetChannel(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
+			fakeClient := mockGetstream.NewStreamClientMock()
+			g := getstream.NewServiceGetStream(fakeClient)
 
-			_, err := c.GetChannel(tt.args.ctx, tt.args.channelID)
+			if tt.name == "Sad case - multiple channels found" {
+				fakeClient.MockQueryChannelsFn = func(ctx context.Context, q *stream.QueryOption, sort ...*stream.SortOption) (*stream.QueryChannelsResponse, error) {
+					return &stream.QueryChannelsResponse{
+						Channels: []*stream.Channel{{ID: gofakeit.UUID()}, {ID: gofakeit.UUID()}},
+					}, nil
+				}
+			}
+
+			if tt.name == "Happy case - retrieve channel by id" {
+				fakeClient.MockQueryChannelsFn = func(ctx context.Context, q *stream.QueryOption, sort ...*stream.SortOption) (*stream.QueryChannelsResponse, error) {
+					return &stream.QueryChannelsResponse{
+						Channels: []*stream.Channel{{ID: gofakeit.UUID()}},
+					}, nil
+				}
+			}
+
+			if tt.name == "Sad case - non existent channel" {
+				fakeClient.MockQueryChannelsFn = func(ctx context.Context, q *stream.QueryOption, sort ...*stream.SortOption) (*stream.QueryChannelsResponse, error) {
+					return &stream.QueryChannelsResponse{
+						Channels: []*stream.Channel{},
+					}, nil
+				}
+			}
+
+			if tt.name == "Sad case - error retrieving channel" {
+				fakeClient.MockQueryChannelsFn = func(ctx context.Context, q *stream.QueryOption, sort ...*stream.SortOption) (*stream.QueryChannelsResponse, error) {
+					return nil, fmt.Errorf("failed to retrieve channel")
+				}
+			}
+
+			_, err := g.GetChannel(tt.args.ctx, tt.args.channelID)
 			if (err != nil) != tt.wantErr {
 				t.Errorf("ChatClient.GetChannel() error = %v, wantErr %v", err, tt.wantErr)
 				return
@@ -263,249 +213,7 @@ func TestChatClient_GetChannel(t *testing.T) {
 
 }
 
-func TestChatClient_RejectInvite(t *testing.T) {
-	ctx := context.Background()
-
-	type args struct {
-		ctx       context.Context
-		userID    string
-		channelID string
-		message   *stream.Message
-	}
-	tests := []struct {
-		name    string
-		args    args
-		want    *stream.Response
-		wantErr bool
-	}{
-		{
-			name: "Happy case",
-			args: args{
-				ctx:       ctx,
-				userID:    userToRejectInviteID,
-				channelID: channelID,
-				message:   nil,
-			},
-			wantErr: false,
-		},
-		{
-			name: "Sad case: non existent user",
-			args: args{
-				ctx:       ctx,
-				userID:    "non-existent-user",
-				channelID: channelID,
-				message:   nil,
-			},
-			wantErr: true,
-		},
-		{
-			name: "Sad case: invalid channel id",
-			args: args{
-				ctx:       ctx,
-				userID:    userToRejectInviteID,
-				channelID: "",
-				message:   nil,
-			},
-			wantErr: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := c.RejectInvite(tt.args.ctx, tt.args.userID, tt.args.channelID, tt.args.message)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("ChatClient.RejectInvite() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if !tt.wantErr && got == nil {
-				t.Errorf("expected a response but got: %v", err)
-			}
-		})
-	}
-}
-
-func TestChatClient_AcceptInvite(t *testing.T) {
-	ctx := context.Background()
-	customInviteMessage := "the user " + userToAcceptInviteName + "accepted the invite"
-	type args struct {
-		ctx       context.Context
-		userID    string
-		channelID string
-		message   *stream.Message
-	}
-	tests := []struct {
-		name    string
-		args    args
-		want    *stream.Response
-		wantErr bool
-	}{
-		{
-			name: "Happy case",
-			args: args{
-				ctx:       ctx,
-				userID:    userToAcceptInviteID,
-				channelID: channelID,
-				message: &stream.Message{
-					Text: customInviteMessage,
-					User: &stream.User{
-						ID:   userToAcceptInviteID,
-						Name: userToAcceptInviteName,
-					},
-				},
-			},
-			wantErr: false,
-		},
-		{
-			name: "Sad case: user does not exist",
-			args: args{
-				ctx:       ctx,
-				userID:    "no-existent-user",
-				channelID: channelID,
-				message:   nil,
-			},
-			wantErr: true,
-		},
-		{
-			name: "Sad case: channel does not exist",
-			args: args{
-				ctx:       ctx,
-				userID:    userToAcceptInviteID,
-				channelID: "no-existent-channel",
-				message:   &stream.Message{Text: customInviteMessage},
-			},
-			wantErr: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := c.AcceptInvite(tt.args.ctx, tt.args.userID, tt.args.channelID, tt.args.message)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("ChatClient.AcceptInvite() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if !tt.wantErr && got == nil {
-				t.Errorf("ChatClient.AcceptInvite() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func TestChatClient_RemoveMembers(t *testing.T) {
-	ctx := context.Background()
-	type args struct {
-		ctx       context.Context
-		channelID string
-		memberIDs []string
-		message   *stream.Message
-	}
-	tests := []struct {
-		name    string
-		args    args
-		want    *stream.Response
-		wantErr bool
-	}{
-		{
-			name: "Happy case",
-			args: args{
-				ctx:       ctx,
-				channelID: channelID,
-				memberIDs: []string{userRemoveFromCommunityID},
-				message:   nil,
-			},
-			wantErr: false,
-		},
-		{
-			name: "Sad case: channel does not exist",
-			args: args{
-				ctx:       ctx,
-				channelID: "no-existent-channel",
-				memberIDs: []string{userRemoveFromCommunityID},
-				message:   nil,
-			},
-			wantErr: true,
-		},
-		{
-			name: "Sad case: non-existent member",
-			args: args{
-				ctx:       ctx,
-				channelID: channelID,
-				memberIDs: []string{"no-existent-member"},
-				message:   nil,
-			},
-			wantErr: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := c.RemoveMembersFromCommunity(tt.args.ctx, tt.args.channelID, tt.args.memberIDs, tt.args.message)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("ChatClient.RemoveMembersFromCommunity() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if !tt.wantErr && got == nil {
-				t.Errorf("expected a response but got: %v", err)
-			}
-		})
-	}
-}
-
-func TestChatClient_DemoteModerators(t *testing.T) {
-	ctx := context.Background()
-
-	type args struct {
-		ctx       context.Context
-		channelID string
-		memberIDs []string
-	}
-	tests := []struct {
-		name    string
-		args    args
-		want    *stream.Response
-		wantErr bool
-	}{
-		{
-			name: "Happy case: demote moderators",
-			args: args{
-				ctx:       ctx,
-				channelID: channelID,
-				memberIDs: []string{moderatorToDemoteID},
-			},
-			wantErr: false,
-		},
-		{
-			name: "Sad case: non-existent channel",
-			args: args{
-				ctx:       ctx,
-				channelID: "no-existent-channel",
-				memberIDs: []string{moderatorToDemoteID},
-			},
-			wantErr: true,
-		},
-		{
-			name: "Sad case: invalid user id",
-			args: args{
-				ctx:       ctx,
-				channelID: channelID,
-				memberIDs: []string{"no-existent-member"},
-			},
-			wantErr: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := c.DemoteModerators(tt.args.ctx, tt.args.channelID, tt.args.memberIDs)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("ChatClient.DemoteModerators() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if !tt.wantErr && got == nil {
-				t.Errorf("expected a response but got: %v", err)
-			}
-		})
-	}
-}
-
 func TestChatClient_RevokeGetStreamUserToken(t *testing.T) {
-	ctx := context.Background()
 
 	type args struct {
 		ctx    context.Context
@@ -518,25 +226,27 @@ func TestChatClient_RevokeGetStreamUserToken(t *testing.T) {
 		wantErr bool
 	}{
 		{
-			name: "Happy case",
+			name: "Happy case: revoke stream token",
 			args: args{
-				ctx:    ctx,
-				userID: userToRevokeGetstreamTokenID,
+				ctx:    context.Background(),
+				userID: gofakeit.UUID(),
 			},
 			wantErr: false,
 		},
-		{
-			name: "Sad case",
-			args: args{
-				ctx:    ctx,
-				userID: "5ea9dc51-a67e-4e5d-aaba-c590c9a66b67",
-			},
-			wantErr: true,
-		},
 	}
+
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			got, err := c.RevokeGetStreamUserToken(tt.args.ctx, tt.args.userID)
+			fakeClient := mockGetstream.NewStreamClientMock()
+			g := getstream.NewServiceGetStream(fakeClient)
+
+			if tt.name == "Happy case: revoke stream token" {
+				fakeClient.MockRevokeUserTokenFn = func(ctx context.Context, userID string, before *time.Time) (*stream.Response, error) {
+					return &stream.Response{}, nil
+				}
+			}
+
+			got, err := g.RevokeGetStreamUserToken(tt.args.ctx, tt.args.userID)
 			if (err != nil) != tt.wantErr {
 				t.Errorf("ChatClient.RevokeGetStreamUserToken() error = %v, wantErr %v", err, tt.wantErr)
 				return
@@ -548,124 +258,7 @@ func TestChatClient_RevokeGetStreamUserToken(t *testing.T) {
 	}
 }
 
-func TestChatClient_BanUser(t *testing.T) {
-	ctx := context.Background()
-	type args struct {
-		ctx            context.Context
-		targetMemberID string
-		bannedBy       string
-		communityID    string
-	}
-	tests := []struct {
-		name    string
-		args    args
-		want    bool
-		wantErr bool
-	}{
-		{
-			name: "Happy Case",
-			args: args{
-				ctx:            ctx,
-				targetMemberID: userToBanID,
-				bannedBy:       defaultModeratorID,
-				communityID:    ch.Channel.ID,
-			},
-			want:    true,
-			wantErr: false,
-		},
-		{
-			name: "Sad Case",
-			args: args{
-				ctx:            ctx,
-				targetMemberID: "",
-				bannedBy:       "",
-				communityID:    "",
-			},
-			want:    false,
-			wantErr: true,
-		},
-		{
-			name: "Sad Case - same targetMemberID",
-			args: args{
-				ctx:            ctx,
-				targetMemberID: userToBanID,
-				bannedBy:       defaultModeratorID,
-				communityID:    channelCID,
-			},
-			want:    false,
-			wantErr: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := c.BanUser(tt.args.ctx, tt.args.targetMemberID, tt.args.bannedBy, tt.args.communityID)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("ChatClient.BanUser() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if got != tt.want {
-				t.Errorf("ChatClient.BanUser() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func TestChatClient_UnBanUser(t *testing.T) {
-	ctx := context.Background()
-
-	type args struct {
-		ctx         context.Context
-		targetID    string
-		communityID string
-	}
-	tests := []struct {
-		name    string
-		args    args
-		want    bool
-		wantErr bool
-	}{
-		{
-			name: "Happy case",
-			args: args{
-				ctx:         ctx,
-				targetID:    userToUnbanID,
-				communityID: channelID,
-			},
-			want:    true,
-			wantErr: false,
-		},
-		{
-			name: "Sad case",
-			args: args{
-				ctx:         ctx,
-				targetID:    "6d743db7-d2ea-4364-a581-b15bad19ada7",
-				communityID: channelID,
-			},
-			want:    false,
-			wantErr: false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := c.UnBanUser(tt.args.ctx, tt.args.targetID, tt.args.communityID)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("ChatClient.UnBanUser() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if got != tt.want {
-				t.Errorf("ChatClient.UnBanUser() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-
-	if err != nil {
-		t.Errorf("ChatClient.DeleteUsers() error = %v", err)
-		return
-	}
-}
-
 func TestChatClient_ListCommunityBannedMembers(t *testing.T) {
-	ctx := context.Background()
 
 	type args struct {
 		ctx         context.Context
@@ -678,17 +271,40 @@ func TestChatClient_ListCommunityBannedMembers(t *testing.T) {
 		wantErr bool
 	}{
 		{
-			name: "Happy case",
+			name: "Happy case: list banned members",
 			args: args{
-				ctx:         ctx,
-				communityID: channelID,
+				ctx:         context.Background(),
+				communityID: gofakeit.UUID(),
 			},
 			wantErr: false,
+		},
+		{
+			name: "sad case: error listing banned members",
+			args: args{
+				ctx:         context.Background(),
+				communityID: gofakeit.UUID(),
+			},
+			wantErr: true,
 		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			_, err := c.ListCommunityBannedMembers(tt.args.ctx, tt.args.communityID)
+			fakeClient := mockGetstream.NewStreamClientMock()
+			g := getstream.NewServiceGetStream(fakeClient)
+
+			if tt.name == "Happy case: list banned members" {
+				fakeClient.MockQueryBannedUsersFn = func(ctx context.Context, q *stream.QueryBannedUsersOptions, sorters ...*stream.SortOption) (*stream.QueryBannedUsersResponse, error) {
+					return &stream.QueryBannedUsersResponse{}, nil
+				}
+			}
+
+			if tt.name == "sad case: error listing banned members" {
+				fakeClient.MockQueryBannedUsersFn = func(ctx context.Context, q *stream.QueryBannedUsersOptions, sorters ...*stream.SortOption) (*stream.QueryBannedUsersResponse, error) {
+					return nil, fmt.Errorf("fail to query banned users")
+				}
+			}
+
+			_, err := g.ListCommunityBannedMembers(tt.args.ctx, tt.args.communityID)
 			if (err != nil) != tt.wantErr {
 				t.Errorf("ChatClient.ListCommunityBannedMembers() error = %v, wantErr %v", err, tt.wantErr)
 				return
@@ -710,20 +326,47 @@ func TestChatClient_UpsertUser(t *testing.T) {
 		wantErr bool
 	}{
 		{
-			name: "Happy case",
+			name: "Happy case: upsert user",
 			args: args{
 				ctx: context.Background(),
 				user: &stream.User{
-					ID:   userToUpsertID,
-					Name: "Test",
+					ID:   gofakeit.UUID(),
+					Name: gofakeit.Name(),
 					Role: "moderator",
 				},
 			},
+			wantErr: false,
+		},
+		{
+			name: "Sad case: fail to upsert user",
+			args: args{
+				ctx: context.Background(),
+				user: &stream.User{
+					ID:   gofakeit.UUID(),
+					Name: gofakeit.Name(),
+					Role: "moderator",
+				},
+			},
+			wantErr: true,
 		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			got, err := c.UpsertUser(tt.args.ctx, tt.args.user)
+			fakeClient := mockGetstream.NewStreamClientMock()
+			g := getstream.NewServiceGetStream(fakeClient)
+
+			if tt.name == "Happy case: upsert user" {
+				fakeClient.MockUpsertUserFn = func(ctx context.Context, user *stream.User) (*stream.UpsertUserResponse, error) {
+					return &stream.UpsertUserResponse{User: &stream.User{ID: gofakeit.UUID()}}, nil
+				}
+			}
+
+			if tt.name == "Sad case: fail to upsert user" {
+				fakeClient.MockUpsertUserFn = func(ctx context.Context, user *stream.User) (*stream.UpsertUserResponse, error) {
+					return nil, fmt.Errorf("failed to upsert user")
+				}
+			}
+			got, err := g.UpsertUser(tt.args.ctx, tt.args.user)
 			if (err != nil) != tt.wantErr {
 				t.Errorf("ChatClient.UpsertUser() error = %v, wantErr %v", err, tt.wantErr)
 				return
@@ -736,7 +379,6 @@ func TestChatClient_UpsertUser(t *testing.T) {
 }
 
 func TestChatClient_DeleteUsers(t *testing.T) {
-	ctx := context.Background()
 
 	type args struct {
 		ctx     context.Context
@@ -750,10 +392,10 @@ func TestChatClient_DeleteUsers(t *testing.T) {
 		wantErr bool
 	}{
 		{
-			name: "Happy case",
+			name: "Happy case: success delete user",
 			args: args{
-				ctx:     ctx,
-				userIDs: []string{userToDeleteID},
+				ctx:     context.Background(),
+				userIDs: []string{gofakeit.UUID()},
 				options: stream.DeleteUserOptions{
 					User:     stream.HardDelete,
 					Messages: stream.HardDelete,
@@ -761,10 +403,38 @@ func TestChatClient_DeleteUsers(t *testing.T) {
 			},
 			wantErr: false,
 		},
+		{
+			name: "Sad Case: error deleting user",
+			args: args{
+				ctx:     context.Background(),
+				userIDs: []string{gofakeit.UUID()},
+				options: stream.DeleteUserOptions{
+					User:     stream.HardDelete,
+					Messages: stream.HardDelete,
+				},
+			},
+			wantErr: true,
+		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			got, err := c.DeleteUsers(tt.args.ctx, tt.args.userIDs, tt.args.options)
+			fakeClient := mockGetstream.NewStreamClientMock()
+			g := getstream.NewServiceGetStream(fakeClient)
+
+			if tt.name == "Happy case: success delete user" {
+				fakeClient.MockDeleteUsersFn = func(ctx context.Context, userIDs []string, options stream.DeleteUserOptions) (*stream.AsyncTaskResponse, error) {
+					return &stream.AsyncTaskResponse{}, nil
+				}
+			}
+
+			if tt.name == "Sad Case: error deleting user" {
+				fakeClient.MockDeleteUsersFn = func(ctx context.Context, userIDs []string, options stream.DeleteUserOptions) (*stream.AsyncTaskResponse, error) {
+					return nil, fmt.Errorf("failed to delete user")
+				}
+
+			}
+
+			got, err := g.DeleteUsers(tt.args.ctx, tt.args.userIDs, tt.args.options)
 			if (err != nil) != tt.wantErr {
 				t.Errorf("ChatClient.DeleteUsers() error = %v, wantErr %v", err, tt.wantErr)
 				return
@@ -777,7 +447,6 @@ func TestChatClient_DeleteUsers(t *testing.T) {
 }
 
 func TestChatClient_ListFlaggedMessages(t *testing.T) {
-	ctx := context.Background()
 
 	type args struct {
 		ctx   context.Context
@@ -790,33 +459,48 @@ func TestChatClient_ListFlaggedMessages(t *testing.T) {
 		wantErr bool
 	}{
 		{
-			name: "Happy case",
+			name: "Happy case: list flagged messages",
 			args: args{
-				ctx: ctx,
+				ctx: context.Background(),
 				input: &stream.QueryOption{
 					Filter: map[string]interface{}{
-						"channel_cid": channelID,
+						"channel_cid": gofakeit.UUID(),
 					},
 				},
 			},
 			wantErr: false,
 		},
 		{
-			name: "Sad case",
+			name: "Sad case: error listing messages",
 			args: args{
-				ctx: ctx,
+				ctx: context.Background(),
 				input: &stream.QueryOption{
 					Filter: map[string]interface{}{
 						"channel_cid": uuid.New().String(),
 					},
 				},
 			},
-			wantErr: false,
+			wantErr: true,
 		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			got, err := c.ListFlaggedMessages(tt.args.ctx, tt.args.input)
+			fakeClient := mockGetstream.NewStreamClientMock()
+			g := getstream.NewServiceGetStream(fakeClient)
+
+			if tt.name == "Happy case: list flagged messages" {
+				fakeClient.MockQueryMessageFlagsFn = func(ctx context.Context, q *stream.QueryOption) (*stream.QueryMessageFlagsResponse, error) {
+					return &stream.QueryMessageFlagsResponse{}, nil
+				}
+			}
+
+			if tt.name == "Sad case: error listing messages" {
+				fakeClient.MockQueryMessageFlagsFn = func(ctx context.Context, q *stream.QueryOption) (*stream.QueryMessageFlagsResponse, error) {
+					return nil, fmt.Errorf("failed to query flagged messages")
+				}
+			}
+
+			got, err := g.ListFlaggedMessages(tt.args.ctx, tt.args.input)
 			if (err != nil) != tt.wantErr {
 				t.Errorf("ChatClient.ListFlaggedMessages() error = %v, wantErr %v", err, tt.wantErr)
 				return
@@ -830,8 +514,6 @@ func TestChatClient_ListFlaggedMessages(t *testing.T) {
 }
 
 func TestChatClient_DeleteMessage(t *testing.T) {
-	ctx := context.Background()
-
 	type args struct {
 		ctx       context.Context
 		messageID string
@@ -843,9 +525,17 @@ func TestChatClient_DeleteMessage(t *testing.T) {
 		wantErr bool
 	}{
 		{
-			name: "Sad case",
+			name: "Happy case: delete a message",
 			args: args{
-				ctx:       ctx,
+				ctx:       context.Background(),
+				messageID: gofakeit.UUID(),
+			},
+			wantErr: false,
+		},
+		{
+			name: "Sad case: fail to delete message",
+			args: args{
+				ctx:       context.Background(),
 				messageID: uuid.New().String(),
 			},
 			wantErr: true,
@@ -853,9 +543,207 @@ func TestChatClient_DeleteMessage(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			got, err := c.DeleteMessage(tt.args.ctx, tt.args.messageID)
+			fakeClient := mockGetstream.NewStreamClientMock()
+			g := getstream.NewServiceGetStream(fakeClient)
+
+			if tt.name == "Happy case: delete a message" {
+				fakeClient.MockDeleteMessageFn = func(ctx context.Context, msgID string) (*stream.Response, error) {
+					return &stream.Response{}, nil
+				}
+			}
+
+			if tt.name == "Sad case: fail to delete message" {
+				fakeClient.MockDeleteMessageFn = func(ctx context.Context, msgID string) (*stream.Response, error) {
+					return nil, fmt.Errorf("failed to delete messages")
+				}
+			}
+
+			got, err := g.DeleteMessage(tt.args.ctx, tt.args.messageID)
 			if (err != nil) != tt.wantErr {
 				t.Errorf("ChatClient.DeleteMessage() error = %v, wantErr %v", err, tt.wantErr)
+				return
+			}
+			if !tt.wantErr && got == nil {
+				t.Errorf("expected a response but got: %v", got)
+				return
+			}
+		})
+	}
+}
+
+func TestChatClient_CreateGetStreamUser(t *testing.T) {
+	type args struct {
+		ctx  context.Context
+		user *stream.User
+	}
+	tests := []struct {
+		name    string
+		args    args
+		wantErr bool
+	}{
+		{
+			name: "Happy case: create user",
+			args: args{
+				ctx: context.Background(),
+				user: &stream.User{
+					ID:   uuid.New().String(),
+					Name: gofakeit.Name(),
+				},
+			},
+			wantErr: false,
+		},
+		{
+			name: "Sad case: fail to upsert user",
+			args: args{
+				ctx: context.Background(),
+				user: &stream.User{
+					ID:   uuid.New().String(),
+					Name: gofakeit.Name(),
+				},
+			},
+			wantErr: true,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			fakeClient := mockGetstream.NewStreamClientMock()
+			g := getstream.NewServiceGetStream(fakeClient)
+
+			if tt.name == "Happy case: create user" {
+				fakeClient.MockUpsertUserFn = func(ctx context.Context, user *stream.User) (*stream.UpsertUserResponse, error) {
+					return &stream.UpsertUserResponse{User: &stream.User{ID: gofakeit.UUID()}}, nil
+				}
+			}
+
+			if tt.name == "Sad case: fail to upsert user" {
+				fakeClient.MockUpsertUserFn = func(ctx context.Context, user *stream.User) (*stream.UpsertUserResponse, error) {
+					return nil, fmt.Errorf("failed to upsert user")
+				}
+			}
+
+			got, err := g.CreateGetStreamUser(tt.args.ctx, tt.args.user)
+			if (err != nil) != tt.wantErr {
+				t.Errorf("ChatClient.CreateGetStreamUser() error = %v, wantErr %v", err, tt.wantErr)
+				return
+			}
+			if !tt.wantErr && got == nil {
+				t.Errorf("expected a response but got: %v", got)
+				return
+			}
+		})
+	}
+}
+
+func TestChatClient_CreateGetStreamUserToken(t *testing.T) {
+	type args struct {
+		ctx    context.Context
+		userID string
+	}
+	tests := []struct {
+		name    string
+		args    args
+		want    string
+		wantErr bool
+	}{
+		{
+			name: "happy Case - successfully generate a user token",
+			args: args{
+				ctx:    context.Background(),
+				userID: gofakeit.UUID(),
+			},
+			want:    "none",
+			wantErr: false,
+		},
+		{
+			name: "sad Case - error generating a user token",
+			args: args{
+				ctx:    context.Background(),
+				userID: gofakeit.UUID(),
+			},
+			want:    "",
+			wantErr: true,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			fakeClient := mockGetstream.NewStreamClientMock()
+			c := getstream.NewServiceGetStream(fakeClient)
+
+			if tt.name == "happy Case - successfully generate a user token" {
+				fakeClient.MockCreateTokenFn = func(userID string, expire time.Time, issuedAt ...time.Time) (string, error) {
+					return "none", nil
+				}
+			}
+
+			if tt.name == "sad Case - error generating a user token" {
+				fakeClient.MockCreateTokenFn = func(userID string, expire time.Time, issuedAt ...time.Time) (string, error) {
+					return "", fmt.Errorf("fail to generate token")
+				}
+			}
+
+			got, err := c.CreateGetStreamUserToken(tt.args.ctx, tt.args.userID)
+			if (err != nil) != tt.wantErr {
+				t.Errorf("ChatClient.CreateGetStreamUserToken() error = %v, wantErr %v", err, tt.wantErr)
+				return
+			}
+			if got != tt.want {
+				t.Errorf("ChatClient.CreateGetStreamUserToken() = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
+
+func TestChatClient_DeleteChannels(t *testing.T) {
+	type args struct {
+		ctx        context.Context
+		chanIDs    []string
+		hardDelete bool
+	}
+	tests := []struct {
+		name    string
+		args    args
+		want    *stream.AsyncTaskResponse
+		wantErr bool
+	}{
+		{
+			name: "happy case: delete channels",
+			args: args{
+				ctx:        context.Background(),
+				chanIDs:    []string{gofakeit.UUID()},
+				hardDelete: true,
+			},
+			wantErr: false,
+		},
+		{
+			name: "sad case: fail to delete channels",
+			args: args{
+				ctx:        context.Background(),
+				chanIDs:    []string{gofakeit.UUID()},
+				hardDelete: true,
+			},
+			wantErr: true,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			fakeClient := mockGetstream.NewStreamClientMock()
+			c := getstream.NewServiceGetStream(fakeClient)
+
+			if tt.name == "happy case: delete channels" {
+				fakeClient.MockDeleteChannelsFn = func(ctx context.Context, cids []string, hardDelete bool) (*stream.AsyncTaskResponse, error) {
+					return &stream.AsyncTaskResponse{}, nil
+				}
+			}
+
+			if tt.name == "sad case: fail to delete channels" {
+				fakeClient.MockDeleteChannelsFn = func(ctx context.Context, cids []string, hardDelete bool) (*stream.AsyncTaskResponse, error) {
+					return nil, fmt.Errorf("fail to delete channels")
+				}
+			}
+
+			got, err := c.DeleteChannels(tt.args.ctx, tt.args.chanIDs, tt.args.hardDelete)
+			if (err != nil) != tt.wantErr {
+				t.Errorf("ChatClient.DeleteChannels() error = %v, wantErr %v", err, tt.wantErr)
 				return
 			}
 			if !tt.wantErr && got == nil {
@@ -878,71 +766,129 @@ func TestChatClient_GetStreamUser(t *testing.T) {
 		wantErr bool
 	}{
 		{
-			name: "happy case: get user by id",
+			name: "happy case: retrieve stream user",
 			args: args{
 				ctx: context.Background(),
-				id:  userToAcceptInviteID,
+				id:  gofakeit.UUID(),
 			},
 			wantErr: false,
 		},
 		{
-			name: "sad case: non existent user",
+			name: "sad case: multiple stream users",
 			args: args{
 				ctx: context.Background(),
-				id:  "user-no-exist",
+				id:  gofakeit.UUID(),
+			},
+			wantErr: true,
+		},
+		{
+			name: "sad case: no stream user",
+			args: args{
+				ctx: context.Background(),
+				id:  gofakeit.UUID(),
+			},
+			wantErr: true,
+		},
+		{
+			name: "sad case: error retrieving stream user",
+			args: args{
+				ctx: context.Background(),
+				id:  gofakeit.UUID(),
 			},
 			wantErr: true,
 		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
+			fakeClient := mockGetstream.NewStreamClientMock()
+			c := getstream.NewServiceGetStream(fakeClient)
+
+			if tt.name == "happy case: retrieve stream user" {
+				fakeClient.MockQueryUsersFn = func(ctx context.Context, q *stream.QueryOption, sorters ...*stream.SortOption) (*stream.QueryUsersResponse, error) {
+					return &stream.QueryUsersResponse{Users: []*stream.User{{ID: gofakeit.UUID()}}}, nil
+				}
+			}
+
+			if tt.name == "sad case: multiple stream users" {
+				fakeClient.MockQueryUsersFn = func(ctx context.Context, q *stream.QueryOption, sorters ...*stream.SortOption) (*stream.QueryUsersResponse, error) {
+					return &stream.QueryUsersResponse{Users: []*stream.User{{ID: gofakeit.UUID()}, {ID: gofakeit.UUID()}}}, nil
+				}
+			}
+
+			if tt.name == "sad case: no stream user" {
+				fakeClient.MockQueryUsersFn = func(ctx context.Context, q *stream.QueryOption, sorters ...*stream.SortOption) (*stream.QueryUsersResponse, error) {
+					return &stream.QueryUsersResponse{Users: []*stream.User{}}, nil
+				}
+			}
+
+			if tt.name == "sad case: error retrieving stream user" {
+				fakeClient.MockQueryUsersFn = func(ctx context.Context, q *stream.QueryOption, sorters ...*stream.SortOption) (*stream.QueryUsersResponse, error) {
+					return nil, fmt.Errorf("failed to query users")
+				}
+			}
+
 			got, err := c.GetStreamUser(tt.args.ctx, tt.args.id)
 			if (err != nil) != tt.wantErr {
 				t.Errorf("ChatClient.GetStreamUser() error = %v, wantErr %v", err, tt.wantErr)
 				return
 			}
 			if !tt.wantErr && got == nil {
-				t.Errorf("ChatClient.GetStreamUser() = %v, want %v", got, tt.want)
+				t.Errorf("expected a response but got: %v", got)
+				return
 			}
 		})
 	}
 }
 
-func TestChatClient_QueryChannelMembers(t *testing.T) {
+func TestChatClient_ValidateGetStreamRequest(t *testing.T) {
 	type args struct {
 		ctx       context.Context
-		channelID string
-		input     *stream.QueryOption
+		body      []byte
+		signature string
 	}
 	tests := []struct {
-		name    string
-		args    args
-		want    *stream.QueryMembersResponse
-		wantErr bool
+		name string
+		args args
+		want bool
 	}{
 		{
-			name: "happy case",
+			name: "happy case: valid stream hook request",
 			args: args{
 				ctx:       context.Background(),
-				channelID: channelID,
-				input: &stream.QueryOption{
-					Filter: map[string]interface{}{
-						"channel_cid": channelID,
-					},
-				},
+				body:      make([]byte, 10),
+				signature: gofakeit.HackerPhrase(),
 			},
+			want: true,
+		},
+		{
+			name: "sad case: invalid stream hook request",
+			args: args{
+				ctx:       context.Background(),
+				body:      make([]byte, 10),
+				signature: gofakeit.HackerPhrase(),
+			},
+			want: false,
 		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
+			fakeClient := mockGetstream.NewStreamClientMock()
+			c := getstream.NewServiceGetStream(fakeClient)
 
-			got, err := c.QueryChannelMembers(tt.args.ctx, tt.args.channelID, tt.args.input)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("ChatClient.QueryChannelMembers() error = %v, wantErr %v", err, tt.wantErr)
-				return
+			if tt.name == "happy case: valid stream hook request" {
+				fakeClient.MockVerifyWebhookFn = func(body, signature []byte) (valid bool) {
+					return true
+				}
 			}
-			if !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("ChatClient.QueryChannelMembers() = %v, want %v", got, tt.want)
+
+			if tt.name == "sad case: invalid stream hook request" {
+				fakeClient.MockVerifyWebhookFn = func(body, signature []byte) (valid bool) {
+					return false
+				}
+			}
+
+			if got := c.ValidateGetStreamRequest(tt.args.ctx, tt.args.body, tt.args.signature); got != tt.want {
+				t.Errorf("ChatClient.ValidateGetStreamRequest() = %v, want %v", got, tt.want)
 			}
 		})
 	}
